requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
cmake==3.18.4.post1
comtypes==1.1.10
Django==3.1.7
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpeg-python==0.2.0
future==0.18.2
google==3.0.0
idna==2.10
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==2.0.3
numpy==1.20.0
opencv-python==4.5.1.48
Pillow==8.1.0
pypiwin32==223
pyttsx3==2.71
pytz==2021.1
pywin32==300
requests==2.25.1
selenium==3.141.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.3
video-to-ascii==1.2.9
xtermcolor==1.3
